# Bug 1526452 - Stop exporting sourcemaps to MC, only support during startmc