--- a/docs/v2-system-addon/1.GETTING_STARTED.md
+++ b/docs/v2-system-addon/1.GETTING_STARTED.md
@@ -81,11 +81,14 @@ mk_add_options MOZ_OBJDIR=./objdir-frontend
 into the `browser/components/newtab` directory inside Mozilla Central.
 3. Build and run Firefox from the `mozilla-central` directory by running `./mach build && ./mach run`.
 
-## Continuous development
+## Continuous development / debugging
 
 Running `npm run startmc` will start a process that watches files in `activity-stream`
 and continuously builds/copies changes to `mozilla-central`. You will
 still need to rebuild Firefox (`./mach build`) if you change `.jsm` files.
+
+**IMPORTANT NOTE**: This task will add inline source maps to help with debugging, which changes the memory footprint.
+Do not use the `startmc` task for profiling or performance testing!
 
 ## Unit Tests