--- a/jar.mn
+++ b/jar.mn
@@ -20,24 +20,12 @@ browser.jar:
   res/activity-stream/data/content/assets/ (./data/content/assets/*)
   res/activity-stream/data/content/tippytop/ (./data/content/tippytop/*)
   res/activity-stream/data/content/activity-stream.bundle.js (./data/content/activity-stream.bundle.js)
-#ifndef RELEASE_OR_BETA
-  res/activity-stream/data/content/activity-stream.bundle.js.map (./data/content/activity-stream.bundle.js.map)
-#endif
 #ifdef XP_MACOSX
   res/activity-stream/css/activity-stream.css (./css/activity-stream-mac.css)
-#ifndef RELEASE_OR_BETA
-  res/activity-stream/css/activity-stream-mac.css.map (./css/activity-stream-mac.css.map)
-#endif
 #elifdef XP_WIN
   res/activity-stream/css/activity-stream.css (./css/activity-stream-windows.css)
-#ifndef RELEASE_OR_BETA
-  res/activity-stream/css/activity-stream-windows.css.map (./css/activity-stream-windows.css.map)
-#endif
 #else
   res/activity-stream/css/activity-stream.css (./css/activity-stream-linux.css)
-#ifndef RELEASE_OR_BETA
-  res/activity-stream/css/activity-stream-linux.css.map (./css/activity-stream-linux.css.map)
-#endif
 #endif
   res/activity-stream/prerendered/static/activity-stream-initial-state.js (./prerendered/static/activity-stream-initial-state.js)
 #ifndef RELEASE_OR_BETA