--- a/webpack.aboutlibrary.config.js
+++ b/webpack.aboutlibrary.config.js
@@ -1,7 +1,7 @@
 const path = require("path");
 const config = require("./webpack.system-addon.config.js");
 const absolute = relPath => path.join(__dirname, relPath);
-module.exports = Object.assign({}, config, {
+module.exports = Object.assign({}, config(), {
   entry: absolute("content-src/aboutlibrary/aboutlibrary.jsx"),
   output: {
     path: absolute("aboutlibrary/content"),