--- a/webpack.system-addon.config.js
+++ b/webpack.system-addon.config.js
@@ -5,14 +5,15 @@ const absolute = relPath => path.join(__dirname, relPath);
 
 const resourcePathRegEx = /^resource:\/\/activity-stream\//;
 
-module.exports = {
+module.exports = (env = {}) => ({
   mode: "none",
   entry: absolute("content-src/activity-stream.jsx"),
   output: {
     path: absolute("data/content"),
     filename: "activity-stream.bundle.js",
   },
-  devtool: "source-map",
+  // TODO: switch to eval-source-map for faster builds. Requires CSP changes
+  devtool: env.development ? "inline-source-map" : false,
   plugins: [new webpack.optimize.ModuleConcatenationPlugin()],
   module: {
     rules: [
@@ -54,4 +55,4 @@ module.exports = {
     "redux": "Redux",
     "react-redux": "ReactRedux",
   },
-};
+});