This patch removes external sourcemaps and retains them only during the startmc, as inline sourcemaps. To test this, try using devtools with `npm run startmc` (you should see line numbers, modules) and `npm run buildmc` (you should not see line numbers, modules).

cc @Mardak 

Notes:
* We don't currently have a clean up task that would remove existing files, we could consider adding that
* An additional patch will be needed to remove the existing files on the MC side