@Mardak 2019-02-11
> We don't currently have a clean up task that would remove existing files, we could consider adding that

Different from `npm run prebuildmc` ?