@Mardak 2019-02-11
> An additional patch will be needed to remove the existing files on the MC side

This should just happen "automatically" as part of the normal export process with `buildmc`:
```
mozilla-central$ git status
	modified:   browser/components/newtab/css/activity-stream-linux.css
	deleted:    browser/components/newtab/css/activity-stream-linux.css.map
	modified:   browser/components/newtab/css/activity-stream-mac.css
	deleted:    browser/components/newtab/css/activity-stream-mac.css.map
	modified:   browser/components/newtab/css/activity-stream-windows.css
	deleted:    browser/components/newtab/css/activity-stream-windows.css.map
	modified:   browser/components/newtab/data/content/activity-stream.bundle.js
	deleted:    browser/components/newtab/data/content/activity-stream.bundle.js.map
```