@dmose 2019-02-12
FWIW, I would expect inlining the source maps will change the memory footprint (and maybe some perf characteristics) of the dev build.  I suspect that doesn't matter, but it might.