@dmose 2019-02-12
> > We don't currently have a clean up task that would remove existing files, we could consider adding that
> 
> Different from `npm run prebuildmc` ?
> 
> Edit: Oh, I see. Existing files from activity-stream side to prevent them from being copied over with buildmc.

If we don't do this, I think we're going to at least want to document in the export checklist that everyone who ever exports probably needs to do this once to inadvertently landing old maps.