@k88hudson 2019-02-12
@dmose Good to note, I'll add that to the comments above the `startmc` task definition; presumably we can just use mozilla-central directly or `buildmc` if we're doing any perf/profiling work.