@dmose 2019-02-12
@k88hudson As far as the perf thing, I think this probably wants to also be documented somewhere in our markdown docs about developing.