@k88hudson 2019-02-14
I have updated the documentation for exports with a notice (see https://docs.google.com/document/d/1olEyxOfoF1C9qi-x8jHCZgynY4czXtmJCx2eKm8Wayc/edit#heading=h.wzgy79ypf8tl) and will sent out a quick note via email / slack when this lands.

As we discussed, I will not implement automation to check for source maps accidentally being included and a bug will be filed only if we see an error show up in a Phabricator patch.