@Mardak 2019-02-15
From my quick testing, it doesn't seem like much of a startmc:webpack regression with this PR:

```
PR/inline
Time: 1226ms
Time: 1265ms
Time: 1362ms
Time: 1396ms
Time: 1479ms

before/separate
Time: 1137ms
Time: 1180ms
Time: 1332ms
Time: 1344ms
Time: 1444ms

eval
Time: 1285ms
Time: 1290ms
Time: 1350ms
Time: 1382ms
Time: 1440ms
```

Compared to before, maybe 30-100ms regression, so 2%-7% regression. Although maybe some people switch between their editor and open firefox builds faster than me. ;)